Rework connection blocking
Previously, message processing was blocked via signalling. When blocked,
only a message that was being waited for could be processed and anything
else was held pending.

Now, we have a connection state. When setting up the connection, any
message that is received is held pending (regardless of whether or not
there's a sync callback for it).