Fix object reference counts
marshalObject is meant to keep a count of the number of references from
that object to any other QObject, as part of the object reference
counting for safe garbage collection.

This was being cleared repeatedly during the marshalling, so its value
would generally be completely wrong and likely lead to dangling
references.